Aggregator for RSS version 2.0 was designed and implemented
Aggregator for RSS version 2.0 was designed and implemented. RSS xml
file was fetched and processed to get the feeds timely. The
implementation consisted of suite of scripts, bash, AWK and Perl. Bash
script was the backbone script, AWK was used for data extraction (text
extraction) and Perl was used to process this data and build a HTML web
page to display the feeds.